requirements.txt
djangorestframework==3.14.0
psycopg2-binary==2.9.5
# requests==2.28.2
